requirements.txt
adafruit-ampy==1.1.0
asgiref==3.7.2
async-timeout==4.0.3
bitarray==2.9.2
bitstring==4.1.4
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.2.7
cryptography==42.0.5
Django==3.2.25
ecdsa==0.18.0
esp==0.2.0
esptool==4.7.0
Flask==2.2.5
Flask-MQTT==1.2.1
idna==3.6
importlib-metadata==6.7.0
intelhex==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.3
linecache2==1.0.0
MarkupSafe==2.1.5
mock==5.1.0
nose==1.3.7
paho-mqtt==1.6.1
pycparser==2.21
pyserial==3.5
python-dotenv==0.21.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
reedsolo==1.7.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
traceback2==1.4.0
typing_extensions==4.7.1
unittest2==1.1.0
uos-activpal==0.2.2
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
